+++ dmemfs/__init__.py
@@ -31,4 +31,4 @@ __all__ = [
     "AsyncMemoryFileSystem",
     "AsyncMemoryFileHandle",
 ]
-__version__ = "0.2.2"
+__version__ = "0.3.0"